LayoutTests/media/modern-media-controls/airplay-button/airplay-button.html

<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;
scheduler.debug = debug;

description("Testing the <code>AirplayButton</code> class.");

const airplayButton = new AirplayButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("airplayButton.element.localName", "button");
shouldBeTrue('airplayButton.element.classList.contains("icon")');
shouldBeTrue('airplayButton.element.classList.contains("airplay")');
shouldBeEqualToString("airplayButton.iconName", `${Icons.Airplay}`);
shouldBe("airplayButton.width", "0");
shouldBe("airplayButton.height", "0");
shouldBe("airplayButton._image", "iconService.imageForIconNameAndLayoutTraits(Icons.Airplay, LayoutTraits.macOS)");
shouldNotBeEqualToString("airplayButton._image.src", "");
shouldBeFalse("airplayButton._image.complete");

// const macOSFullscreenAirplayButton = new AirplayButton({ layoutTraits: LayoutTraits.macOS | LayoutTraits.Fullscreen });
// const iOSAirplayButton = new AirplayButton({ layoutTraits: LayoutTraits.iOS });

airplayButton._image.addEventListener("load", event => {
    shouldNotBe("airplayButton.width", "0");
    shouldNotBe("airplayButton.height", "0");
    shouldBeTrue("airplayButton.needsLayout");
    shouldBeTrue("scheduler.hasScheduledLayoutCallbacks");
    shouldBe("scheduler._layoutCallbacks.size", "1");
    shouldNotBe("scheduler._frameID", "-1");
    shouldBeTrue("dirtyNodes.has(airplayButton)");
    scheduler.frameWillFire = function() {
        debug("frameWillFire()");
        shouldBeTrue("dirtyNodes.has(airplayButton)");
        shouldBeTrue("airplayButton.needsLayout");
    }
    scheduler.frameDidFire = function() {
        debug("frameDidFire()");
        shouldBeFalse("dirtyNodes.has(airplayButton)");
        shouldBeFalse("airplayButton.needsLayout");
    }
    shouldBecomeEqual('airplayButton.element.style.webkitMaskImage.includes("macOS/airplay@")', "true", finishJSTest);
    // shouldBecomeEqual('airplayButton.element.style.webkitMaskImage.includes("macOS/airplay@")', "true", () => {
    //     shouldBecomeEqual('macOSFullscreenAirplayButton.element.style.webkitMaskImage.includes("macOS/airplay-fullscreen@")', "true", () => {
    //         shouldBecomeEqual('iOSAirplayButton.element.style.webkitMaskImage.includes("iOS/airplay@")', "true", finishJSTest);
    //     });
    // });
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
